+++ docs/changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.0-7] - 2022-12-13
+
+### Added
+
+- Added `BulmaFormField`, `BulmaFormControl`, and `BulmaFormControlGroup` components
+
+### Changed
+
+- Typescript defintions to include comments/documentation of props, types, functions, etc.
+
 ## [0.0.0-6.1] - 2022-12-13
 
 ### Fixed